% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{check.eig.sign}
\alias{check.eig.sign}
\title{Modified version of algorithm from
Bro, Acar, and Kolda ''Resolving the sign ambiguity in the singular value
decomposition'' (2008), J Chemometrics (22) 135-140.}
\usage{
check.eig.sign(A, B, X, Y)
}
\arguments{
\item{A}{Numeric matrix. Left canonical vectors (loadings).}

\item{B}{Numeric matrix. Right canonical vectors (loadings).}

\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}
}
\description{
Modified version of algorithm from
Bro, Acar, and Kolda ''Resolving the sign ambiguity in the singular value
decomposition'' (2008), J Chemometrics (22) 135-140.
}
