% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{fcca}
\alias{fcca}
\title{Fit the FCCA model.}
\usage{
fcca(
  X,
  Y,
  ndim = 1,
  lambda1 = 0,
  lambda2 = 0,
  gamma1 = 0,
  gamma2 = 0,
  V = NULL,
  standx = c("sd", "none"),
  standy = c("sd", "none"),
  svd.tol = 1e-12,
  maxiter = 1000,
  verbose = FALSE
)
}
\arguments{
\item{X}{An n by p numeric matrix. Note: PLINK data is currently not supported.}

\item{Y}{An n by k numeric matrix.}

\item{ndim}{Integer. Positive number of canonical vectors to compute.}

\item{lambda1}{Numeric vector. Non-negative L1 penalty on canonical vectors of X.}

\item{lambda2}{Numeric vector. Non-negative L1 penalty on canonical vectors of Y.}

\item{gamma1}{Numeric vector. Non-negative L2 penalty on X.}

\item{gamma2}{Numeric vector. Non-negative L2 penalty on Y.}

\item{V}{Numeric matrix. Initial estimate of the left singular
vector of  X'Y.}

\item{standx}{Character. One of "sd" (empricial standard deviation) or "none".}

\item{standy}{Character. One of "sd" (empricial standard deviation) or "none".}

\item{svd.tol}{Numeric. Tolerance under which to truncate singular values of X and Y.}

\item{maxiter}{Integer. Maximum number of iterations.}

\item{verbose}{Logical. Whether to print more information.}
}
\value{
an \code{fcca} object with the following components:

\describe{
   \item{ndim:} Number of dimensions.
   \item{lambda1:} Vector of lambda1 penalties.
   \item{lambda2:} Vector of lambda2 penalties.
   \item{gamma1:} Vector of gamma1 penalties.
   \item{gamma2:} Vector of gamma2 penalties.
   \item{U:} Matrix of left singular vectors
   \item{a:} Matrix of canonical vectors for X.
   \item{V:} Matrix of right singular vectors
   \item{b:} Matrix of canonical vectors for Y.
   \item{Px:} Matrix of canonical coordinates, X * a
   \item{Py:} Matrix of canonical coordinates, Y * b
   \item{r:} Canonical correlations, diag(cor(Px, Py))
   \item{converged:} whether the optimisation has converged or not
}
}
\description{
Fit the FCCA model.
}
\details{
Standardisation is done once for X and Y
}
