\name{flashpca-package}
\alias{flashpca-package}
\alias{flashpca}
\docType{package}
\title{
R interface to flashpca
}
\description{
An R interface to flashpca for fast principal component analysis.
}
\details{
\tabular{ll}{
Package: \tab flashpca\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-30\cr
License: \tab GPL (>= 3)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Gad Abraham

Maintainer: Gad Abraham <gad.abraham@unimelb.edu.au>
}
\references{
Abraham G, Inouye M (2014) Fast Principal Component Analysis of
Large-Scale Genome-Wide Data. PLoS ONE 9(4): e93766. doi:10.1371/journal.pone.0093766
}
\keyword{ package }
\seealso{
}
\examples{
}
