% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm3_chr1.R
\docType{data}
\name{hm3.chr1}
\alias{hm3.chr1}
\title{Subset of chromosome 1 from HapMap3}
\format{A list with elements
\describe{
\item{bed}{A matrix of allele dosage 0, 1, 2}
\item{bim}{The PLINK bim file}
\item{fam}{The PLINK fam file}
\item{population}{The relationship file indicating which population each
person is from}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/hapmap/genotypes/}
}
\usage{
hm3.chr1
}
\description{
A dataset containing (LD-thinned)
SNP genotypes from chromosome 1 in 957 individuals
from HapMap3.
}
\keyword{datasets}

