% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcca.R
\name{pcca}
\alias{pcca}
\title{Principal component canonical correlation analysis (PCCA)}
\usage{
pcca(
  X,
  Y,
  ndim = NULL,
  kx = 3,
  ky = 3,
  standx = c("sd", "none"),
  standy = c("sd", "none"),
  check_sign = TRUE,
  svd_tol = 1e-12
)
}
\arguments{
\item{X}{An n by p numeric matrix. Note: PLINK data is currently not supported.}

\item{Y}{An n by k numeric matrix.}

\item{ndim}{Integer. Positive number of canonical vectors to compute.}

\item{kx}{Numeric. Number of principal components to compute for X.}

\item{ky}{Numeric. Number of principal components to compute for Y.}

\item{standx}{Character. One of "sd" (empricial standard deviation) or "none".}

\item{standy}{Character. One of "sd" (empricial standard deviation) or "none".}

\item{check_sign}{Logical. Whether to check and correct possible sign flips 
of the singular vectors within the cross-validation.}

\item{svd_tol}{Numeric. Tolerance under which to truncate singular values of X and Y.}

\item{nfolds}{Integer. Number of cross-validation folds.}

\item{folds}{Integer. The fold identifiers. Overrides the `nfolds'
parameter.}
}
\description{
Principal component canonical correlation analysis (PCCA)
}
