% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{plot.cv.scca}
\alias{plot.cv.scca}
\title{Plot the results of SCCA cross-validation}
\usage{
\method{plot}{cv.scca}(x, dim = NULL, ...)
}
\arguments{
\item{x}{An object of class "cv.scca"}

\item{dim}{Integer. Which dimension to plot (all will be plotted by
default).}

\item{...}{Other arguments that will be passed to \code{plot}.}
}
\description{
Plot the results of SCCA cross-validation
}
\details{
Plots the cross-validated Pearson correlation, as a function of the
   number of non-zero entries in the canonical vectors U (on the x-axis),
   with a separate curve separately for each lambda2 penalty (which affects
   the non-zero entries in the canonical vector V. For example, if
   \code{cv.scca} was run with 5x lambda1 values and 7x lambda2 values,
   there will be 7 separate curves each with 5 points on the x-axis.
}
\examples{

#######################
## HapMap3 chr1 example
data(hm3.chr1)
X <- scale2(hm3.chr1$bed)
n <- nrow(X)
m <- ncol(X)
k <- 5
B <- matrix(rnorm(m * k), m, k)
Y <- X \%*\% B + rnorm(n * k)

r <- cv.scca(X, Y, standx="sd", standy="sd", nfolds=3, ndim=2,
   lambda1=seq(1e-3, 1e-1, length=10),
   lambda2=seq(1e-4, 0.5, length=3))

par(mfrow=c(1, 2))
plot(r, dim=1)
plot(r, dim=2)

}

