% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcca.R
\name{print.cv.fcca}
\alias{print.cv.fcca}
\title{Prints a cv.fcca object}
\usage{
\method{print}{cv.fcca}(x, ...)
}
\arguments{
\item{x}{A cv.fcca object to be printed}

\item{...}{Ignored}
}
\description{
Prints a cv.fcca object
}
