% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcca.R
\name{print.cv.pcca}
\alias{print.cv.pcca}
\title{Prints a cv.pcca object}
\usage{
\method{print}{cv.pcca}(x, ...)
}
\arguments{
\item{x}{A cv.pcca object to be printed}

\item{...}{Ignored}
}
\description{
Prints a cv.pcca object
}
