% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{print.fcca}
\alias{print.fcca}
\title{Prints an fcca object}
\usage{
\method{print}{fcca}(x, ...)
}
\arguments{
\item{x}{An fcca object to be printed}

\item{...}{Ignored}
}
\description{
Prints an fcca object
}
