% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashpca.R
\name{print.flashpca}
\alias{print.flashpca}
\title{Prints a flashpca object}
\usage{
\method{print}{flashpca}(x, ...)
}
\arguments{
\item{x}{A flashpca object to be printed}

\item{...}{Ignored.}
}
\description{
Prints a flashpca object
}

