% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcca.R
\name{print.optim.cv.fcca}
\alias{print.optim.cv.fcca}
\title{Prints an optim.cv.fcca object}
\usage{
\method{print}{optim.cv.fcca}(x, ...)
}
\arguments{
\item{x}{An optim.cv.fcca object to be printed}

\item{...}{Ignored}
}
\description{
Prints an optim.cv.fcca object
}
