% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcca.R
\name{print.pcca}
\alias{print.pcca}
\title{Prints a pcca object}
\usage{
\method{print}{pcca}(x, ...)
}
\arguments{
\item{x}{A pcca object to be printed}

\item{...}{Ignored}
}
\description{
Prints a pcca object
}
