% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{print.scca}
\alias{print.scca}
\title{Prints an scca object}
\usage{
\method{print}{scca}(x, ...)
}
\arguments{
\item{x}{An scca object to be printed}

\item{...}{Ignored}
}
\description{
Prints an scca object
}
