% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucca.R
\name{print.ucca}
\alias{print.ucca}
\title{Prints a UCCA object}
\usage{
\method{print}{ucca}(x, ...)
}
\arguments{
\item{x}{A flashpca object to be printed}

\item{...}{Ignored}
}
\description{
Prints a UCCA object
}

