% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\docType{data}
\name{rna}
\alias{rna}
\title{Gene expression data}
\format{

}
\source{
\url{http://www.ebi.ac.uk/arrayexpress/files/E-TABM-1036/E-TABM-1036.processed.1.zip}
}
\usage{
rna
}
\description{
A dataset containing gene expression levels
}
\keyword{datasets}
