% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{Standardise a genotype matrix X.}
\usage{
scale2(X, type = c("2", "1"), impute = TRUE)
}
\arguments{
\item{X}{A genotype matrix in dosage encoding 0, 1, 2.}

\item{type}{A character string indicating which type of standard deviation
to use.}

\item{impute}{Logical. Whether to impute missing values to zero (=mean
after standardisation).}
}
\description{
Standardise a genotype matrix X.
}
\details{
{
type 1 (old Eigensoft way, Price 2006)
   mean = 2p (p is MAF)
   sd = sqrt(p (1 - p))

type 2: (new Eigensoft way)
   mean = 2p (p is MAF)
   sd = sqrt(2 * p (1 - p))

Missing genotypes ('NA') are imputed to the mean (=0).
}
}

