% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucca.R
\name{ucca}
\alias{ucca}
\title{Performs univariate canonical correlation analysis, i.e., ANOVA of all
phenotypes on each SNP.}
\usage{
ucca(X, Y, standx = c("binom2", "binom", "sd", "center", "none"),
  standy = c("binom2", "binom", "sd", "center", "none"), check_geno = TRUE,
  check_fam = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{An n by p numeric matrix, or a character string pointing to a
PLINK dataset}

\item{Y}{An n by k numeric matrix of phenotypes.}

\item{standx}{Character. One of "binom" (zero mean, unit variance
where variance is p * (1 - p), for SNP data), "binom2" (zero mean, unit
variance where variance is 2 * p * (1 - p),
"sd" (zero-mean, unit Gaussian variance), "center" (zero mean), or "none". Note
that if you use "binom" for non-SNP data you may get garbage. Note that
currently the same standardisation is applied to both X and Y. If you
require different standardisations, it's best to standardise yourself
and then choose standx="none".}

\item{standy}{Character. Stanardisation of Y.}

\item{check_geno}{Logical. Whether to explicitly check if the matrices
X and Y contain values other than {0, 1, 2}, when standx or standy is one 
"binom" or "binom2". This can
be set to FALSE if you are sure your matrices only contain these values
(only matters when using "binom"/"binom2").}

\item{check_fam}{Logical. Whether to check that the number of row in 
the PLINK fam file (if X is a character string) matches the number of
rows in the eigenvectors.}

\item{verbose}{Logical.}
}
\value{
\code{ucca} returns a list containing the following components:

\describe{  
   \item{result:}{A numeric matrix of the association results, in the original ordering.}
   \item{npheno:}{The number of phenotypes used.}
   \item{nsnps:}{The number of SNPs used.}
}
}
\description{
Performs univariate canonical correlation analysis, i.e., ANOVA of all
phenotypes on each SNP.
}
\details{
This is an efficient implementation of Ferreira and Purcell's
plink.multivariate test for association between multiple phenotypes and one SNP at a time.

This test is equivalent to the F-test for a linear regression of each SNP on the
phenotypes \code{lm(SNP ~ pheno1 + pheno2 + ...)},
hence the number of phenotypes that can be tested is limited by
the sample size (the model is not penalised).

By default, \code{ucca} performs mean-imputation of missing values.
}
\examples{

data(hm3.chr1)
n <- nrow(hm3.chr1$bed)
p <- ncol(hm3.chr1$bed)
k <- 10
X <- scale2(hm3.chr1$bed)
By <- matrix(rnorm(p * k), p, k)
Y <- scale(X \%*\% By + rnorm(n * k))

# Call on a numeric matrix X;
# X has already been standardised
s1 <- ucca(X, Y, standx="none", standy="sd")

# Call on a PLINK dataset
bedf <- gsub("\\\\.bed", "",
    system.file("extdata", "data_chr1.bed", package="flashpcaR"))
s2 <- ucca(bedf, Y, standx="binom2", standy="sd")

head(s1$result)
head(s2$result)

# Same result (to within numerical tolerance)
mean((s1$result[, "Fstat"] - s2$result[, "Fstat"])^2)

}
\references{
M. A. R. Ferreira and S. M. Purcell (2009) _A multivariate test
of association_, Bioinformatics 25(1) 132-133
}

