% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{validate.rank}
\alias{validate.rank}
\title{Estimates the effective number of dimensions of a matrix X, based on
reconstruction error from singular value decomposition.}
\usage{
validate.rank(X, maxdim = 20, test.prop = 0.1, const = 0)
}
\arguments{
\item{X}{numeric matrix.}

\item{maxdim}{Integer. How many dimensions to try.}

\item{test.prop}{numeric. What proportion of the data to reserve for
evaluating the reconstruction error on.}

\item{const}{numeric. What value to set the test observations to.}
}
\value{
a data.frame with number of dimensions and reconstruction error
(mse).
}
\description{
Estimates the effective number of dimensions of a matrix X, based on
reconstruction error from singular value decomposition.
}
\details{
A certain proportion of the entries of X is set to a constant
value. We then run singular value decomposition (SVD) on the matrix
and for each rank k compute the reconstruction mean squared error between
the predicted values \eqn{U_{1:k} D_{1:k} V_{1:k}^T} and the true values at
at these missing positions. The rank k with the lowest mean squared error
is chosen as the effective rank for whitening.

This is basically one step in the imputation method from
http://alexhwilliams.info/itsneuronalblog/2018/02/26/crossval
}
\examples{

#######################
## HapMap3 chr1 example
data(hm3.chr1)
X <- scale2(hm3.chr1$bed)

r <- validate.rank(X, maxdim=50)
plot(mse ~ dim, data=r)

}
