% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{whiten}
\alias{whiten}
\title{Performs a low-rank whitening of a matrix X.}
\usage{
whiten(X, ndim = 20, keep.svd = TRUE, stand = "none")
}
\arguments{
\item{X}{numeric matrix.}

\item{ndim}{Integer. How many dimensions to use in the whitening.}

\item{keep.svd}{Logical. Whether to keep the singular value decomposition
results.}

\item{stand}{character. Which standardisation to use on columns of X.}
}
\description{
Performs a low-rank whitening of a matrix X.
}
\details{
A k-rank singular value decomposition of X is done
\deqn{X = U_{1:k} D_{1:k} V_{1:k}^T} 
The
}
\examples{

data(hm3.chr1)
X <- scale2(hm3.chr1$bed)
Xw <- whiten(X, ndim=10)

names(attr(Xw, "svd"))

}
