% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{scca}
\alias{scca}
\title{Performs sparse canonical correlation analysis.}
\usage{
scca(X, Y, lambda1 = 0, lambda2 = 0, standx = c("binom2", "binom", "sd",
  "center", "none"), standy = c("binom2", "binom", "sd", "center", "none"),
  ndim = 10, maxiter = 1000, tol = 1e-04, seed = 1L, verbose = FALSE,
  num_threads = 1, mem = c("low", "high"), check_geno = TRUE,
  check_fam = TRUE, V = NULL, block_size = 500, simplify = TRUE)
}
\arguments{
\item{X}{An n by p numeric matrix, or a character string pointing to a
PLINK dataset.}

\item{Y}{An n by k numeric matrix}

\item{lambda1}{Numeric. Non-negative L1 penalty on canonical vectors of X.}

\item{lambda2}{Numeric. Non-negative L1 penalty on canonical vectors of Y.}

\item{standx}{Character. One of "binom" (zero mean, unit variance
where variance is p * (1 - p), for SNP data), "binom2" (zero mean, unit
variance where variance is 2 * p * (1 - p),
"sd" (zero-mean, unit Gaussian variance), "center" (zero mean), or "none". Note
that if you use "binom" for non-SNP data you may get garbage. Note that
currently the same standardization is applied to both X and Y. If you
require different standardizations, it's best to standardise yourself
and then choose standx="none". When X is a PLINK dataset name, standx must be one of "binom2" 
or "binom".}

\item{standy}{Character. Stanardisation of Y.}

\item{ndim}{Integer. Positive number of canonical vectors to compute.}

\item{maxiter}{Integer. Positive, maximum number of iterations to perform.}

\item{tol}{Numeric. Tolerance for convergence.}

\item{seed}{Integer. Random seed for initialisation of CCA iterations.}

\item{verbose}{Logical.}

\item{num_threads}{Integer. Number of OpenMP threads to
use (only supported under Linux and on Mac if compiled with GCC).}

\item{mem}{Character. One of "low" or "high", where "low"
doesn't explicitly compute X^T Y, and "high" does.
This is useful for large X and Y.}

\item{check_geno}{Logical. Whether to explicitly check if the matrices
X and Y contain values other than {0, 1, 2}, when standx/standy="binom"/"binom2". This can
be set to FALSE if you are sure your matrices only contain these values
(only matters when using "binom" or "binom2").}

\item{check_fam}{Logical. Whether to check that the number of rows in 
the PLINK fam file (if X is a character string) matches the number of
rows in the eigenvectors.}

\item{V}{Numeric. A vector to initialise "v" in SCCA iterations. By
default, it will be a vector of normally distributed variates.}

\item{block_size}{Integer. Size of blocks for reading PLINK data.}

\item{simplify}{Logical. Whether to return a single \code{scca} object or a
list when only one model is fitted.}
}
\value{
\code{scca} returns a list containing the following components:

\describe{  
   \item{U:}{Top ndim left canonical vectors of X^T Y.}
   \item{V:}{Top ndim right canonical vectors of X^T Y.}
   \item{d:}{Top ndim canonical covariances, i.e., diag(cov(X U, Y V)). 
 Note that we don't divide by n-1.}
   \item{Px:}{X * U.}
   \item{Py:}{Y * V.}
}
}
\description{
Performs sparse canonical correlation analysis.
}
\examples{

#######################
## HapMap3 chr1 example
data(hm3.chr1)
X <- scale2(hm3.chr1$bed)
n <- nrow(X)
m <- ncol(X)
k <- 10
B <- matrix(rnorm(m * k), m, k)
Y <- X \%*\% B + rnorm(n * k)

s <- scca(X, Y, lambda1=1e-2, lambda2=1e-2, ndim=5,
  standx="none", standy="sd")

## The canonical correlations
diag(cor(s$Px, s$Py))

}

